Add some output; spt use new lua state

+++ src/spectrum.h
@@ -33,7 +33,7 @@ enum var_type{
 
 struct string{
     char *s;
-    size_t l;
+    uint64_t l;
 };
 
 struct number{
@@ -71,10 +71,10 @@ struct sp_thread{
     record_t *record;
     record_t *record_tail;
 
-    size_t records_num;
-    size_t records_num_nomatch;
-    size_t records_num_errmatch;
-    size_t records_num_droped;
+    uint64_t records_num;
+    uint64_t records_num_nomatch;
+    uint64_t records_num_errmatch;
+    uint64_t records_num_droped;
 
     iterm_t *logs;
